Demo.Mmose.Core.Test/ServerEngine/World/Zone/ZoneTest.cs
using Demo.Mmose.Core.World.Zone;
using Microsoft.VisualStudio.TestTools.UnitTesting;
using Demo.Mmose.Core.World;
using Demo.Mmose.Core.Network;
using Demo.Mmose.Core.Common;

namespace Demo.Mmose.Core.Test
{
    /// <summary>
    ///这是 ZoneTest 的测试类，旨在
    ///包含所有 ZoneTest 单元测试
    ///</summary>
    [TestClass()]
    public class ZoneTest
    {
        private TestContext testContextInstance;
        /// <summary>
        ///获取或设置测试上下文，上下文提供
        ///有关当前测试运行及其功能的信息。
        ///</summary>
        public TestContext TestContext
        {
            get { return testContextInstance; }
            set { testContextInstance = value; }
        }

        #region 附加测试属性
        // 
        //编写测试时，还可使用以下属性:
        //
        //使用 ClassInitialize 在运行类中的第一个测试前先运行代码
        //[ClassInitialize()]
        //public static void MyClassInitialize(TestContext testContext)
        //{
        //}
        //
        //使用 ClassCleanup 在运行完类中的所有测试后再运行代码
        //[ClassCleanup()]
        //public static void MyClassCleanup()
        //{
        //}
        //
        //使用 TestInitialize 在运行每个测试前先运行代码
        //[TestInitialize()]
        //public void MyTestInitialize()
        //{
        //}
        //
        //使用 TestCleanup 在运行完每个测试后运行代码
        //[TestCleanup()]
        //public void MyTestCleanup()
        //{
        //}
        //
        #endregion

        /// <summary>
        ///ZonePacketHandlers 的测试
        ///</summary>
        public void ZonePacketHandlersTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone<T> target = new Zone<T>(); // TODO: 初始化为适当的值
            ZonePacketHandlers actual;
            actual = target.ZonePacketHandlers;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void ZonePacketHandlersTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ZonePacketHandlersTestHelper<T>()。" );
        }

        /// <summary>
        ///World 的测试
        ///</summary>
        public void WorldTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone<T> target = new Zone<T>(); // TODO: 初始化为适当的值
            T actual;
            actual = target.World;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void WorldTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 WorldTestHelper<T>()。" );
        }

        /// <summary>
        ///ProcessorBuffers 的测试
        ///</summary>
        public void ProcessorBuffersTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone_Accessor<T> target = new Zone_Accessor<T>(); // TODO: 初始化为适当的值
            BufferPool actual;
            actual = target.ProcessorBuffers;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void ProcessorBuffersTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ProcessorBuffersTestHelper<T>()。" );
        }

        /// <summary>
        ///LogInId 的测试
        ///</summary>
        public void LogInIdTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone<T> target = new Zone<T>(); // TODO: 初始化为适当的值
            Serial actual;
            actual = target.LogInId;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void LogInIdTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 LogInIdTestHelper<T>()。" );
        }

        /// <summary>
        ///IsLoginZoneCluster 的测试
        ///</summary>
        public void IsLoginZoneClusterTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone<T> target = new Zone<T>(); // TODO: 初始化为适当的值
            bool actual;
            actual = target.IsLoginZoneCluster;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void IsLoginZoneClusterTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 IsLoginZoneClusterTestHelper<T>()。" );
        }

        /// <summary>
        ///ConfigZone 的测试
        ///</summary>
        public void ConfigZoneTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone<T> target = new Zone<T>(); // TODO: 初始化为适当的值
            ConfigZone actual;
            actual = target.ConfigZone;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void ConfigZoneTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ConfigZoneTestHelper<T>()。" );
        }

        /// <summary>
        ///ZoneInitOnceServer 的测试
        ///</summary>
        public void ZoneInitOnceServerTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone_Accessor<T> target = new Zone_Accessor<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            BaseWorldEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.ZoneInitOnceServer( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void ZoneInitOnceServerTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ZoneInitOnceServerTestHelper<T>()。" );
        }

        /// <summary>
        ///ZoneExitWorld 的测试
        ///</summary>
        public void ZoneExitWorldTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone_Accessor<T> target = new Zone_Accessor<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            BaseWorldEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.ZoneExitWorld( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void ZoneExitWorldTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ZoneExitWorldTestHelper<T>()。" );
        }

        /// <summary>
        ///UnRegToZoneCluster 的测试
        ///</summary>
        public void UnRegToZoneClusterTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone<T> target = new Zone<T>(); // TODO: 初始化为适当的值
            NetState netState = null; // TODO: 初始化为适当的值
            Zone<T>.Option option = new Zone<T>.Option(); // TODO: 初始化为适当的值
            bool expected = false; // TODO: 初始化为适当的值
            bool actual;
            actual = target.UnRegToZoneCluster( netState, option );
            Assert.AreEqual( expected, actual );
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void UnRegToZoneClusterTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 UnRegToZoneClusterTestHelper<T>()。" );
        }

        /// <summary>
        ///RegToZoneCluster 的测试
        ///</summary>
        public void RegToZoneClusterTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone<T> target = new Zone<T>(); // TODO: 初始化为适当的值
            NetState netState = null; // TODO: 初始化为适当的值
            Zone<T>.Option option = new Zone<T>.Option(); // TODO: 初始化为适当的值
            bool expected = false; // TODO: 初始化为适当的值
            bool actual;
            actual = target.RegToZoneCluster( netState, option );
            Assert.AreEqual( expected, actual );
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void RegToZoneClusterTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 RegToZoneClusterTestHelper<T>()。" );
        }

        /// <summary>
        ///OnLoginZoneCluster 的测试
        ///</summary>
        public void OnLoginZoneClusterTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone_Accessor<T> target = new Zone_Accessor<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            NetStateConnectEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.OnLoginZoneCluster( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void OnLoginZoneClusterTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 OnLoginZoneClusterTestHelper<T>()。" );
        }

        /// <summary>
        ///LogoutZone 的测试
        ///</summary>
        public void LogoutZoneTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone<T> target = new Zone<T>(); // TODO: 初始化为适当的值
            target.LogoutZone();
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        public void LogoutZoneTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 LogoutZoneTestHelper<T>()。" );
        }

        /// <summary>
        ///ListenerProcessReceive 的测试
        ///</summary>
        public void ListenerProcessReceiveTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone_Accessor<T> target = new Zone_Accessor<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            NetStateEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.ListenerProcessReceive( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void ListenerProcessReceiveTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ListenerProcessReceiveTestHelper<T>()。" );
        }

        /// <summary>
        ///ListenerInitNetState 的测试
        ///</summary>
        public void ListenerInitNetStateTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone_Accessor<T> target = new Zone_Accessor<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            NetStateInitEventArgs netStateInit = null; // TODO: 初始化为适当的值
            target.ListenerInitNetState( sender, netStateInit );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void ListenerInitNetStateTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ListenerInitNetStateTestHelper<T>()。" );
        }

        /// <summary>
        ///InitZone 的测试
        ///</summary>
        public void InitZoneTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone<T> target = new Zone<T>(); // TODO: 初始化为适当的值
            ConfigZone configZone = null; // TODO: 初始化为适当的值
            target.InitZone( configZone );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        public void InitZoneTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 InitZoneTestHelper<T>()。" );
        }

        /// <summary>
        ///GetPacketLength 的测试
        ///</summary>
        public void GetPacketLengthTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone_Accessor<T> target = new Zone_Accessor<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            PacketLengthInfoEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.GetPacketLength( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void GetPacketLengthTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 GetPacketLengthTestHelper<T>()。" );
        }

        /// <summary>
        ///GetPacketID 的测试
        ///</summary>
        public void GetPacketIDTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone_Accessor<T> target = new Zone_Accessor<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            PacketIdInfoEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.GetPacketID( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void GetPacketIDTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 GetPacketIDTestHelper<T>()。" );
        }

        /// <summary>
        ///ConnectToZoneClusterServer 的测试
        ///</summary>
        public void ConnectToZoneClusterServerTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone_Accessor<T> target = new Zone_Accessor<T>(); // TODO: 初始化为适当的值
            bool expected = false; // TODO: 初始化为适当的值
            bool actual;
            actual = target.ConnectToZoneClusterServer();
            Assert.AreEqual( expected, actual );
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void ConnectToZoneClusterServerTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ConnectToZoneClusterServerTestHelper<T>()。" );
        }

        /// <summary>
        ///ConnecterProcessReceive 的测试
        ///</summary>
        public void ConnecterProcessReceiveTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone_Accessor<T> target = new Zone_Accessor<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            NetStateEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.ConnecterProcessReceive( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void ConnecterProcessReceiveTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ConnecterProcessReceiveTestHelper<T>()。" );
        }

        /// <summary>
        ///ConnecterIniNetState 的测试
        ///</summary>
        public void ConnecterIniNetStateTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone_Accessor<T> target = new Zone_Accessor<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            NetStateInitEventArgs netStateInit = null; // TODO: 初始化为适当的值
            target.ConnecterIniNetState( sender, netStateInit );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void ConnecterIniNetStateTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ConnecterIniNetStateTestHelper<T>()。" );
        }

        /// <summary>
        ///Zone`1 构造函数 的测试
        ///</summary>
        public void ZoneConstructorTestHelper<T>()
            where T : BaseWorld, new()
        {
            Zone<T> target = new Zone<T>();
            Assert.Inconclusive( "TODO: 实现用来验证目标的代码" );
        }

        [TestMethod()]
        public void ZoneConstructorTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ZoneConstructorTestHelper<T>()。" );
        }
    }
}
